Unidad 3

+++ unidad3/Ennumerados/Ennumerados/Program.cs
@@ -17,7 +17,7 @@ namespace Ennumerados
 
         static void Main(string[] args)
         {
-        Console.WriteLine((int) DiasSemana.Lunes);
+        Console.WriteLine( DiasSemana.Lunes);
         Console.WriteLine((int) DiasSemana.Viernes);
         }
     }
